+++ packages/utils/src/types.ts
@@ -1,4 +1,10 @@
-export type Wallet = { name: string; hidden?: boolean; options?: Record<string, any>; connectorName?: string }
+import type { AbstractConnector } from '@web3-react/abstract-connector'
+
+export type Wallet = {
+  name: string
+  hidden?: boolean
+  connector: AbstractConnector
+}
 
 export type TxHistoryFetcher<Tx = any, P = any, O = any> = (opts: {
   address: string