+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rainbow-me/kit-hooks
 
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @rainbow-me/kit-utils@0.0.1
+
 ## 0.0.0
 
 ### Patch Changes