+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rainbow-me/kit-utils
 
+## 0.0.1
+
+### Patch Changes
+
+- make the user import connectors instead of a library doing it
+
 ## 0.0.0
 
 ### Patch Changes